+++ spe-ed-solver/visualisation/src/main/java/visualisation/ColorGradient.java
@@ -3,7 +3,7 @@ package visualisation;
 import java.awt.Color;
 
 /**
- * Color Gradient for generating RGB values for float values between 0 and 1.
+ * Color Gradient for generating RGB values for double values between 0 and 1.
  */
 public enum ColorGradient {
 
@@ -12,10 +12,10 @@ public enum ColorGradient {
 	 */
 	FIRE {
 		@Override
-		protected int applyUnbounded(final float value) {
-			final int rValue = (int) (0xFF * Math.max(0, Math.min(1 / 3f, value)) * 3);
-			final int gValue = (int) (0xFF * Math.max(0, Math.min(1 / 3f, value - 0.33)) * 3);
-			final int bValue = (int) (0xFF * Math.max(0, Math.min(1 / 3f, value - 0.66)) * 3);
+		protected int applyUnbounded(final double value) {
+			final int rValue = (int) (0xFF * Math.max(0, Math.min(1 / 3, value)) * 3);
+			final int gValue = (int) (0xFF * Math.max(0, Math.min(1 / 3, value - 0.33)) * 3);
+			final int bValue = (int) (0xFF * Math.max(0, Math.min(1 / 3, value - 0.66)) * 3);
 			final int rgbValue = (rValue << 16) | (gValue << 8) | bValue;
 			return rgbValue;
 		}
@@ -25,8 +25,8 @@ public enum ColorGradient {
 	 */
 	RAINBOW {
 		@Override
-		protected int applyUnbounded(final float value) {
-			return Color.HSBtoRGB(value, 1, 1);
+		protected int applyUnbounded(final double value) {
+			return Color.HSBtoRGB((float) value, 1, 1);
 		}
 	};
 
@@ -37,17 +37,17 @@ public enum ColorGradient {
 	 * @param value must be between 0 and 1
 	 * @return RGB value as int (0 - 0xFFFFFF)
 	 */
-	protected abstract int applyUnbounded(final float value);
+	protected abstract int applyUnbounded(final double value);
 
 	/**
 	 * Clipping value to range between 0 and 1 and applying RGB gradient
 	 * transformation.
 	 * 
-	 * @param value the float value to get the RGB value for
+	 * @param value the double value to get the RGB value for
 	 * @return RGB value as int (0 - 0xFFFFFF)
 	 */
-	public final int apply(final float value) {
-		float boundedValue = Math.max(0, Math.min(1, value));
+	public final int apply(final double value) {
+		double boundedValue = Math.max(0, Math.min(1, value));
 		return this.applyUnbounded(boundedValue);
 	}
 }
